# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/completions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to create a goal completion",
                "parameters": [
                    {
                        "description": "body params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.CreateGoalCompletionBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/db.GoalCompletion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to delete a goal completion",
                "parameters": [
                    {
                        "description": "body params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.DeleteGoalCompletionBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.DeleteGoalCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    }
                }
            }
        },
        "/goals": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to create a goal",
                "parameters": [
                    {
                        "description": "body params",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.CreateGoalBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/routes.CreateGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    }
                }
            }
        },
        "/pending-goals": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to get pending goals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.PendingGoal"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    }
                }
            }
        },
        "/summary": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Endpoint to get week summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.WeekSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "db.GoalCompletion": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "goalId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "db.SummaryGoal": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "goalId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "db.WeekSummary": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "goalsPerDay": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/db.SummaryGoal"
                        }
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "routes.CreateGoalBody": {
            "type": "object",
            "required": [
                "desiredWeeklyFrequency",
                "title"
            ],
            "properties": {
                "desiredWeeklyFrequency": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "routes.CreateGoalCompletionBody": {
            "type": "object",
            "properties": {
                "goalId": {
                    "type": "string"
                }
            }
        },
        "routes.CreateGoalResponse": {
            "type": "object",
            "properties": {
                "desiredWeeklyFrequency": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "routes.DeleteGoalCompletionBody": {
            "type": "object",
            "properties": {
                "completionId": {
                    "type": "string"
                },
                "goalId": {
                    "type": "string"
                }
            }
        },
        "routes.DeleteGoalCompletionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "routes.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "routes.PendingGoal": {
            "type": "object",
            "properties": {
                "completionCount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "desiredWeeklyFrequency": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}